requirements.txt
Flask >= 1.1.0
numpy >= 1.20.1
Pillow >= 8.1.0
